helm for index DB. Pass DB path to perl scripts not just project name.
Index DB now uses BTREE.

Added a helm-source for the indexed database (helm-vps.el). The first
word must be a prefix of at least fire chars. Type space and a
word (repeatedly) to grep the initial grep output. The first word is a
prefix and the next are substrings "filters".

The path to the DB files is passed to the perl files. This way the vps
dir is not hardcoded in the perl scripts.